requirements.txt
fastapi[all]==0.104.1
uvicorn>=0.18.1
pymongo>=4.1.1
pyjwt>=2.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator>=1.2.1